reimplement prescriber to load all on rank 0
