+++ workflows/prognostic_c48_run/prepare_config.py
@@ -21,7 +21,7 @@ from runtime.diagnostics.fortran import file_configs_to_namelist_settings
 from runtime.steppers.nudging import NudgingConfig
 from runtime.config import UserConfig
 from runtime.steppers.machine_learning import MachineLearningConfig
-from runtime.steppers.prephysics import PrescriberConfig
+from runtime.steppers.prescriber import PrescriberConfig
 
 
 logger = logging.getLogger(__name__)
